refactor: consolidate transport integration tests into a unified module
### CHANGES

- Consolidate HTTP and SSE transport tests into one file.
- Introduce `TransportTestBase` to remove duplicated test code.
- Delete separate HTTP and SSE transport test files.
- Refactor core MCP integration tests for improved clarity.
- Rename test classes and methods for better consistency.
- Centralize transport-related CLI tests in the new module.
