+++ tests/integration/test_mcp_integration.py
@@ -1,7 +1,7 @@
-"""Integration tests for the Fabric MCP Server.
+"""Integration tests for core MCP functionality.
 
-These tests verify the end-to-end functionality of the MCP server,
-including the protocol interactions and tool execution.
+These tests verify the core MCP server functionality, tool registration,
+and protocol interactions without focusing on specific transport types.
 """
 
 import logging
@@ -14,16 +14,14 @@ from unittest.mock import Mock, patch
 
 import httpx
 import pytest
-from anyio import WouldBlock
-from fastmcp import FastMCP
 
 from fabric_mcp import __version__
 from fabric_mcp.core import FabricMCP
 
 
 @pytest.mark.integration
-class TestFabricMCPIntegration:
-    """Integration tests for the complete Fabric MCP Server."""
+class TestFabricMCPCore:
+    """Integration tests for core Fabric MCP Server functionality."""
 
     @pytest.fixture
     def server(self):
@@ -47,16 +45,37 @@ class TestFabricMCPIntegration:
             },
         }
 
-    @pytest.mark.asyncio
-    async def test_mcp_tool_discovery(self, server: FabricMCP):
-        """Test that MCP tools are properly discoverable."""
-        # This would test the list_tools() functionality when implemented
-        # For now, we test that the server has the expected tools registered
+    def test_server_initialization_and_configuration(self, server: FabricMCP):
+        """Test server initialization and configuration."""
+        assert server.log_level == "DEBUG"
+        assert server.mcp.name.startswith("Fabric MCP v")
+        assert hasattr(server, "logger")
         assert hasattr(server, "mcp")
         assert server.mcp is not None
 
+    def test_tool_registration_and_discovery(self, server: FabricMCP):
+        """Test that MCP tools are properly registered and discoverable."""
+        # Check that tools are registered
+        tools = getattr(server, "_FabricMCP__tools", [])
+        assert len(tools) == 6
+
+        # Verify each tool is callable
+        for tool in tools:
+            assert callable(tool)
+
+        # Test specific tool functionality
+        list_patterns_tool = tools[0]
+        result: list[str] = list_patterns_tool()
+        assert isinstance(result, list)
+        assert len(result) == 3
+
+        pattern_details_tool = tools[1]
+        result = pattern_details_tool("test_pattern")
+        assert isinstance(result, dict)
+        assert "name" in result
+
     @pytest.mark.asyncio
-    async def test_fabric_list_patterns_integration(
+    async def test_fabric_list_patterns_with_mocked_api(
         self, server: FabricMCP, mock_fabric_api_response: Mock
     ):
         """Test the fabric_list_patterns tool with mocked API calls."""
@@ -68,17 +87,15 @@ class TestFabricMCPIntegration:
             mock_client.return_value.get.return_value = mock_response
 
             # Execute the tool (currently returns hardcoded values)
-            # In the future, this would make actual API calls
             tools = getattr(server, "_FabricMCP__tools", [])
-            if tools:
-                list_patterns_tool = tools[0]
-                result: list[Callable[..., Any]] = list_patterns_tool()
+            list_patterns_tool = tools[0]
+            result: list[Callable[..., Any]] = list_patterns_tool()
 
-                assert isinstance(result, list)
-                assert len(result) > 0
+            assert isinstance(result, list)
+            assert len(result) > 0
 
     @pytest.mark.asyncio
-    async def test_fabric_pattern_details_integration(
+    async def test_fabric_pattern_details_with_mocked_api(
         self, server: FabricMCP, mock_fabric_api_response: Mock
     ):
         """Test the fabric_pattern_details tool with mocked API calls."""
@@ -93,15 +110,14 @@ class TestFabricMCPIntegration:
 
             # Execute the tool (currently returns hardcoded values)
             tools = getattr(server, "_FabricMCP__tools", [])
-            if len(tools) > 1:
-                pattern_details_tool = tools[1]
-                result = pattern_details_tool("analyze_claims")
+            pattern_details_tool = tools[1]
+            result = pattern_details_tool("analyze_claims")
 
-                assert isinstance(result, dict)
-                assert "name" in result
+            assert isinstance(result, dict)
+            assert "name" in result
 
     @pytest.mark.asyncio
-    async def test_fabric_run_pattern_integration(
+    async def test_fabric_run_pattern_with_mocked_api(
         self, server: FabricMCP, mock_fabric_api_response: Mock
     ):
         """Test the fabric_run_pattern tool with mocked API calls."""
@@ -116,16 +132,15 @@ class TestFabricMCPIntegration:
 
             # Execute the tool (currently returns hardcoded values)
             tools = getattr(server, "_FabricMCP__tools", [])
-            if len(tools) > 2:
-                run_pattern_tool = tools[2]
-                result = run_pattern_tool("analyze_claims", "Test input text")
+            run_pattern_tool = tools[2]
+            result = run_pattern_tool("analyze_claims", "Test input text")
 
-                assert isinstance(result, dict)
-                assert "output_format" in result
-                assert "output_text" in result
+            assert isinstance(result, dict)
+            assert "output_format" in result
+            assert "output_text" in result
 
     @pytest.mark.asyncio
-    async def test_server_error_handling_with_fabric_api_down(self, server: FabricMCP):
+    async def test_error_handling_with_fabric_api_down(self, server: FabricMCP):
         """Test error handling when Fabric API is unavailable."""
         with patch("httpx.Client") as mock_client:
             # Simulate connection error
@@ -134,18 +149,14 @@ class TestFabricMCPIntegration:
             )
 
             # For now, tools return hardcoded values, but in future they should
-            # handle errors
+            # handle errors gracefully
             tools = getattr(server, "_FabricMCP__tools", [])
-            if tools:
-                list_patterns_tool = tools[0]
-                # Current implementation returns hardcoded values, so this passes
-                # In future implementation, this should handle the connection
-                # error gracefully
-                result = list_patterns_tool()
-                assert isinstance(result, list)
+            list_patterns_tool = tools[0]
+            result = list_patterns_tool()
+            assert isinstance(result, list)
 
     @pytest.mark.asyncio
-    async def test_server_error_handling_with_fabric_api_error(self, server: FabricMCP):
+    async def test_error_handling_with_fabric_api_error(self, server: FabricMCP):
         """Test error handling when Fabric API returns errors."""
         with patch("httpx.Client") as mock_client:
             # Simulate HTTP error
@@ -159,14 +170,11 @@ class TestFabricMCPIntegration:
             mock_client.return_value.get.return_value = mock_response
 
             # For now, tools return hardcoded values, but in future they should
-            # handle errors
+            # handle errors gracefully
             tools = getattr(server, "_FabricMCP__tools", [])
-            if tools:
-                list_patterns_tool = tools[0]
-                # Current implementation returns hardcoded values, so this passes
-                # In future implementation, this should handle the HTTP error gracefully
-                result = list_patterns_tool()
-                assert isinstance(result, list)
+            list_patterns_tool = tools[0]
+            result = list_patterns_tool()
+            assert isinstance(result, list)
 
     def test_server_stdio_integration(self, server: FabricMCP):
         """Test the stdio method integration with mocked MCP run."""
@@ -174,113 +182,45 @@ class TestFabricMCPIntegration:
             server.stdio()
             mock_run.assert_called_once()
 
-    def test_server_configuration_integration(self, server: FabricMCP):
-        """Test server configuration and initialization."""
-        assert server.log_level == "DEBUG"
-        assert server.mcp.name.startswith("Fabric MCP v")
-        assert hasattr(server, "logger")
-
-
-class TestFabricMCPProtocol:
-    """Test MCP protocol-specific functionality."""
-
-    @pytest.fixture
-    def server(self):
-        """Create a FabricMCP server instance for testing."""
-        return FabricMCP(log_level="INFO")
-
-    def test_mcp_server_inheritance(self, server: FabricMCP):
-        """Test that FabricMCP properly inherits from FastMCP."""
-
-        assert isinstance(server, FastMCP)
-        assert hasattr(server, "tool")
-        assert hasattr(server, "run")
-
-    @pytest.mark.asyncio
-    async def test_mcp_tool_registration_integration(self, server: FabricMCP):
-        """Test that tools are properly registered with the MCP framework."""
-        # This tests the integration between our tools and the FastMCP framework
-        # The actual tool registration happens in __init__
-        tools = getattr(server, "_FabricMCP__tools", [])
-        assert len(tools) == 6
-
-        # Verify each tool is callable
-        for tool in tools:
-            assert callable(tool)
-
-
-class TestFabricMCPErrorScenarios:
-    """Test error scenarios and edge cases."""
-
-    @pytest.fixture
-    def server(self):
-        """Create a FabricMCP server instance for testing."""
-        return FabricMCP(log_level="ERROR")
-
-    def test_server_graceful_shutdown_on_keyboard_interrupt(
+    def test_server_graceful_shutdown_scenarios(
         self, server: FabricMCP, caplog: pytest.LogCaptureFixture
     ):
-        """Test graceful shutdown on KeyboardInterrupt."""
-
+        """Test graceful shutdown on various interrupt signals."""
         with caplog.at_level(logging.INFO):
+            # Test KeyboardInterrupt
             with patch.object(server.mcp, "run", side_effect=KeyboardInterrupt):
                 server.stdio()
 
-            # Check log messages
-            assert "Server stopped by user." in caplog.text
-
-    def test_server_handles_multiple_shutdown_signals(
-        self, server: FabricMCP, caplog: pytest.LogCaptureFixture
-    ):
-        """Test that server handles various shutdown signals properly."""
-
-        with caplog.at_level(logging.INFO):
             # Test CancelledError
             with patch.object(server.mcp, "run", side_effect=CancelledError):
                 server.stdio()
 
-            # Test WouldBlock
-            with patch.object(server.mcp, "run", side_effect=WouldBlock):
-                server.stdio()
-
-            # Both should result in graceful shutdown messages
-            assert caplog.text.count("Server stopped by user.") >= 2
-
-
-@pytest.mark.integration
-class TestEndToEndScenarios:
-    """End-to-end integration test scenarios."""
-
-    @pytest.fixture
-    def server(self):
-        """Create a FabricMCP server instance for testing."""
-        return FabricMCP(log_level="INFO")
+        # Should have at least one graceful shutdown message
+        assert "Server stopped by user." in caplog.text
 
     @pytest.mark.asyncio
     async def test_complete_pattern_workflow(self, server: FabricMCP):
         """Test a complete workflow: list patterns -> get details -> run pattern."""
         tools = getattr(server, "_FabricMCP__tools", [])
 
-        if len(tools) >= 3:
-            # Step 1: List patterns
-            list_patterns_tool = tools[0]
-            patterns: list[Callable[..., Any]] = list_patterns_tool()
-            assert isinstance(patterns, list)
-            assert len(patterns) > 0
+        # Step 1: List patterns
+        list_patterns_tool = tools[0]
+        patterns: list[str] = list_patterns_tool()
+        assert isinstance(patterns, list)
+        assert len(patterns) > 0
 
-            # Step 2: Get pattern details
-            pattern_details_tool = tools[1]
-            if patterns:
-                details = pattern_details_tool(patterns[0])
-                assert isinstance(details, dict)
-                assert "name" in details
+        # Step 2: Get pattern details
+        pattern_details_tool = tools[1]
+        details = pattern_details_tool("test_pattern")
+        assert isinstance(details, dict)
+        assert "name" in details
 
-            # Step 3: Run pattern
-            run_pattern_tool = tools[2]
-            result = run_pattern_tool(patterns[0], "Test input")
-            assert isinstance(result, dict)
-            assert "output_format" in result
-            assert "output_text" in result
+        # Step 3: Run pattern
+        run_pattern_tool = tools[2]
+        result = run_pattern_tool("test_pattern", "Test input")
+        assert isinstance(result, dict)
+        assert "output_format" in result
+        assert "output_text" in result
 
     def test_server_lifecycle(self, server: FabricMCP):
         """Test complete server lifecycle: init -> configure -> run -> shutdown."""
@@ -289,7 +229,7 @@ class TestEndToEndScenarios:
         assert hasattr(server, "mcp")
 
         # Test configuration
-        assert server.log_level == "INFO"
+        assert server.log_level == "DEBUG"
 
         # Test run with immediate shutdown
         with patch.object(server.mcp, "run", side_effect=KeyboardInterrupt):
@@ -300,8 +240,8 @@ class TestEndToEndScenarios:
 
 
 @pytest.mark.integration
-class TestFabricMCPEndToEnd:
-    """End-to-end integration tests that execute the fabric-mcp script."""
+class TestFabricMCPCLI:
+    """End-to-end integration tests for the fabric-mcp CLI."""
 
     def test_version_flag(self):
         """Test that fabric-mcp --version returns the correct version."""
@@ -331,7 +271,7 @@ class TestFabricMCPEndToEnd:
         assert "--transport" in result.stdout
         assert "--log-level" in result.stdout
 
-    def test_no_args_shows_help(self):
+    def test_no_args_shows_missing_transport_error(self):
         """Test that running fabric-mcp with no args errors with missing transport."""
         result = subprocess.run(
             [sys.executable, "-m", "fabric_mcp.cli"],
